requirements.txt
flask
selenium
linkedin-scraper
openai
reportlab
webdriver-manager
python-dotenv
